remove unused classes and records

+++ CogLog.Persistence/Repos/TopicRepo.cs
@@ -23,10 +23,14 @@ public class TopicRepo(AppDbContext ctx) : BaseRepo<Topic>(ctx), ITopicRepo
         ;
     }
 
-    public async Task DeleteTopicAsync(Topic topic)
+    public async Task DeleteTopicAsync(int id)
     {
-        _ctx.Topics.Remove(topic);
-        await _ctx.SaveChangesAsync();
+        var topic = await _ctx.Topics.FindAsync(id);
+        if (topic != null)
+        {
+            _ctx.Topics.Remove(topic);
+            await _ctx.SaveChangesAsync();
+        }
     }
 
     public async Task<TopicDetailsDto?> GetTopicDetailsAsync(int id)
